requirements.txt
RPi.GPIO==0.6.3
rpi-ws281x==3.0.6